work in progress: main.py refactoring
Added two column layout (as per original design)

Splitting up main.py into main, hello_window, editor
