+++ config/settings.json
@@ -1,0 +1,3 @@
+{
+    "is_signed_in": false
+}
